feat: pre-fill inquiry form from service detail page CTAs
Split /kontakt page into a server component (reads ?service= query param)
and a ContactChannels client component (handles copy-to-clipboard state).
Both CTA buttons on service detail pages now pass the service slug so the
InquiryForm selector is pre-selected on arrival.

Closes #13
